Basic untested registration endpoint

+++ userDB.sql
@@ -1,17 +1,11 @@
-IF (EXISTS userDB)
-BEGIN
-    DROP DATABASE userDB
-END
+DROP DATABASE userDB if EXISTS userDB
 
 CREATE DATABASE userDB
 
 -- create table to contain user information
-IF (NOT EXISTS userInfo)
-BEGIN
-    CREATE TABLE userInfo (
-        uuID VARCHAR(255)
-        username VARCHAR(30)
-        -- salted pw + salt value using 2a format
-        saltedPass BINARY(60)
-    )
-END
+CREATE TABLE userInfo (
+    uuID VARCHAR(255)
+    username VARCHAR(30)
+    -- salted pw + salt value using 2a format
+    saltedPass BINARY(60)
+) if NOT EXISTS userInfo
